Update 1.3.4 for Minecraft 1.21.11
Updated to work with Minecraft 1.21.11 and to use Mojang's mappings. The Bedrock Pickaxe will now check for any block in the "bedrock_like" block tag when preventing you from breaking a block at the bottom of the world (so, by default, bedrock or any bedrock-based block added by the mod can't be broken at the bottom-most Y-level unless you're above a different block or air that's at said Y-level).